third_party/WebKit/LayoutTests/fast/css3-text/css3-text-align-last/getComputedStyle/getComputedStyle-text-align-last-inherited-expected.txt

This test checks that the value of text-align-last is properly inherited to the child.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Value of ancestor is 'start':
PASS window.getComputedStyle(ancestor).textAlignLast is 'start'
PASS window.getComputedStyle(child).textAlignLast is 'start'


Value of ancestor is 'end':
PASS window.getComputedStyle(ancestor).textAlignLast is 'end'
PASS window.getComputedStyle(child).textAlignLast is 'end'


Value of ancestor is 'left':
PASS window.getComputedStyle(ancestor).textAlignLast is 'left'
PASS window.getComputedStyle(child).textAlignLast is 'left'


Value of ancestor is 'right':
PASS window.getComputedStyle(ancestor).textAlignLast is 'right'
PASS window.getComputedStyle(child).textAlignLast is 'right'


Value of ancestor is 'center':
PASS window.getComputedStyle(ancestor).textAlignLast is 'center'
PASS window.getComputedStyle(child).textAlignLast is 'center'


Value of ancestor is 'justify':
PASS window.getComputedStyle(ancestor).textAlignLast is 'justify'
PASS window.getComputedStyle(child).textAlignLast is 'justify'


Value of ancestor is 'auto':
PASS window.getComputedStyle(ancestor).textAlignLast is 'auto'
PASS window.getComputedStyle(child).textAlignLast is 'auto'


Value of ancestor is 'start, while child is 'end':
PASS window.getComputedStyle(ancestor).textAlignLast is 'start'
PASS window.getComputedStyle(child).textAlignLast is 'end'


Value of ancestor is 'left, while child is 'right':
PASS window.getComputedStyle(ancestor).textAlignLast is 'left'
PASS window.getComputedStyle(child).textAlignLast is 'right'


PASS successfullyParsed is true

TEST COMPLETE

hello world
